Fix House structure

+++ src/main/scala/ru/ifmo/ctddev/ml/hw2/House.scala
@@ -4,14 +4,14 @@ import java.io.{File, FileReader}
 
 import breeze.io.CSVReader
 
-case class House(area: Double, rooms: Int)
+case class House(area: Int, rooms: Int)
 
-case class HouseWithPrice(house: House, price: Double)
+case class HouseWithPrice(house: House, price: Int)
 
 object HouseWithPrice {
   def parseData(csvFile: File): Seq[HouseWithPrice] = {
     CSVReader.read(new FileReader(csvFile)).map {
-      case Seq(area, rooms, price) => HouseWithPrice(House(area.toDouble, rooms.toInt), price.toDouble)
+      case Seq(area, rooms, price) => HouseWithPrice(House(area.toInt, rooms.toInt), price.toInt)
       case _ => throw new IllegalArgumentException("CSV is malformed")
     }
   }
